Add HeadlessDashboard utility class
Jira: TNT-418
