+++ libs/sdk-ui-dashboard/src/model/headlessDashboard/index.ts
@@ -1,0 +1,3 @@
+// (C) 2022 GoodData Corporation
+
+export { HeadlessDashboard, HeadlessDashboardConfig, MonitoredAction } from "./HeadlessDashboard";
