+++ libs/sdk-ui-dashboard/src/model/index.ts
@@ -90,3 +90,4 @@ export {
     DashboardEventEvalFn,
 } from "./eventHandlers/eventHandler";
 export { newDrillToSameDashboardHandler } from "./eventHandlers/drillToSameDashboardHandlerFactory";
+export * from "./headlessDashboard";
