+++ libs/sdk-ui-dashboard/src/model/store/index.ts
@@ -172,4 +172,12 @@ export { uiActions } from "./ui";
 
 export { queryAndWaitFor } from "./_infra/queryAndWaitFor";
 export { dispatchAndWaitFor } from "./_infra/dispatchAndWaitFor";
-export { DashboardStore } from "./dashboardStore";
+export {
+    IDashboardQueryService,
+    QueryCache,
+    QueryActions,
+    QueryCacheEntry,
+    QueryCacheEntryResult,
+    AllQueryCacheReducers,
+    QueryCacheReducer,
+} from "./_infra/queryService";
